Hack glslang to accept extensions used in Falcor shaders.
The main things being added here are:
- rudimentary support for GL_ARB_bindless_textures, especially conversion of a `sampler2d` to a `uvec2`
- rudimentary support for the `intptr_t` type introduced in GL_NV_shader_buffer_load
- basic support to parse, and then ignore default values on parameters (may not be needed)

None of these are actually implemented in a proper working fashion, but they at least aspire to not just blindly ignore invalid code.

More serious support for, e.g., the pointer types introduced by shader_buffer_load would require much more significant changes.

All changes attempt to identify themselves with `BEGIN FALCOR` and `END FALCOR`, in order to make merging with future changes to mainline glslang a bit easier.

Note that the `glslang_tab.cpp*` files are automatically generated by bison, so the diffs are not very useful. The version of bison I used is older than the one that the mainline uses, though, so there is a possibility that this could lead to behavior differences.