Checkpoint initial work on cross-compiler.
This adds a new tool to the glslang project (rather than trying to add yet more bells and whistles to the `glslangValidator` project). The basic approach is to take a single GLSL file as input, parse it using the glslang library, and then start to "pretty-print" it as soemthing approaching HLSL code (that last part is *not* meaningfully in place right now).

This is currently building as a stand-alone executable, but I suppose we will want to switch it to a library project relatively soon.

This code is not production-ready in the least; I am currently just working on getting things up and limping and we can focus on polish once we have confirmed that the approach will meet our needs.