Add missing `getAsLoopNode`
The glslang AST/IR type `TIntermNode` has a bunch of virtual functions for dynamic-casting it to the different sub-cases, but is missing a case for `TIntermLoop`. This change adds a `getAsLoopNode()` function.

Note that the "right" way to traverse the AST/IR in glslang is supposed to be using their traverser framework, but that seems to only be a (sort of) good idea if you are lowering to an ISA-like target, and not for purposes like pretty-printing. Also, from looking at the SPIR-V output logic, there seems to be a some fiddly assumptions about order of visitation and side effects.

All told, going with a more ordinary traversal seems to make sense, and this function is needed in order to implement that (modulo just *guessing* that any node which doesn't match the other cases must be a loop).