Basic support for renaming symbols.
There are two basic features here, supported with some shared constructs in the code:

- We track a set of reserved words (currently not properly populated), and if we try to declare a variable or struct field that uses one of those words as its name, we rename it.

- We track a set of names used in each scope, and if we try to re-use the same name in a given scope, we add a numeric suffix to generate a uique name.

Currently these rules really only apply to local variables, and struct fields. Function and struct names aren't changed, because we want people to be able to refer to such declarations normally in code that consumes the translation result.