Emit `cbuffer` decls (more) correctly.
The original code was just emitting them as they were encountered, which obviously wasn't right, and was also using the glslang `getName()` on the uniform block decl instead of the `getTypeName()` on its type. Little did I know that the block might not have an explicit name (and so would get a messy auto-generated symbol), and that the "type name" was actually what an HLSL programmer would have thought of as the block name. It actually makes some kind of sense in retrospect...

The other subtle thing is that glslang always represents access to a uniform in a block as an explicit "dot operator": `block.field`, while HLSL always just uses the implicit form `field`. Rather than kludge up the code for emitting expressions to detect this special case (which honestly wouldn't be *too* bad), I went ahead and generated the HLSL cbuffer in an idiomatic form:

    cbuffer Foo
    {
        struct
        {
            // ... actual block members go here
        } Foo;
    };

This means that the block name for reflection purposes is `Foo`, and the shader code can still refer to members with `Foo.field`.

Gotchas:

- Okay, that clever thing I'm doing with the block name is going to mess up any attempt to enumerate uniforms because there will be this extra prefix. We can probably work around that, but realistically it is better for me to fix the translator by kluding the expression rules in the way I just described...

- I am still not doing anything about cases where we might have sampler/resource types inside a uniform block (using bindless), so they will translate to sort of broken HLSL.

- I haven't done anything to translate over explicit binding stuff (and I'm not sure how it will work since some of our explicit bindings are out of the range supported by HLSL...).